requirements.txt
arrow==1.3.0
asgiref==3.8.1
attrs==22.2.0
binaryornot==0.4.4
build==0.10.0
CacheControl==0.12.11
certifi==2022.12.7
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.1.0
cleo==2.0.1
click==8.1.7
cookiecutter==2.6.0
crashtest==0.4.1
cryptography==39.0.2
distlib==0.3.6
Django==5.1.4
django-browser-reload==1.17.0
django-tailwind==3.8.0
django-unfold==0.43.0
dulwich==0.21.3
et_xmlfile==2.0.0
filelock==3.9.0
html5lib==1.1
idna==3.4
importlib-metadata==6.0.0
installer==0.6.0
jaraco.classes==3.2.3
jeepney==0.8.0
Jinja2==3.1.4
jsonschema==4.17.3
keyring==23.13.1
lockfile==0.12.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
more-itertools==9.1.0
msgpack==1.0.5
numpy==2.1.3
openpyxl==3.1.5
packaging==23.0
pandas==2.2.3
pexpect==4.8.0
pillow==11.0.0
pipenv==2023.2.18
pkginfo==1.9.6
platformdirs==2.6.2
poetry==1.4.0
poetry-core==1.5.1
poetry-plugin-export==1.3.0
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.18.0
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
rapidfuzz==2.15.1
requests==2.28.2
requests-toolbelt==0.10.1
rich==13.9.4
SecretStorage==3.3.3
shellingham==1.5.0.post1
six==1.17.0
sqlparse==0.5.2
text-unidecode==1.3
tomlkit==0.11.6
trove-classifiers==2023.3.9
types-python-dateutil==2.9.0.20241206
tzdata==2024.2
urllib3==1.26.15
virtualenv==20.20.0
virtualenv-clone==0.5.7
webencodings==0.5.1
zipp==3.15.0
